+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Colorize and improve failing tests output.
 - Fix `go` with a generic function: `go test<string>(c, 'abcd')`.
 - Add comptime `x := $embed_file('v.png') println(x.len) println(ptr_str(x.data()))`, for embedding files into binaries.
+- Advanced vdoc search on mobile layout.
 
 ## V 0.2.1
 *30 Dec 2020*